apps/scheduler/serializers.py
from rest_framework import serializers
from .models import Event, Calendar

class EventSerializer(serializers.ModelSerializer):
    class Meta:
        model = Event
        fields = ['id', 'title', 'description', 'start_time', 'end_time']

class CalendarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Calendar
        fields = ['id', 'name', 'user', 'created_at', 'updated_at']