requirements.txt
Django>=4.2
psycopg2-binary>=2.8
django-environ
python-dotenv
django-cors-headers
gunicorn==20.1.0
djangorestframework>=3.14